apps/api/pyproject.toml
[project]
name = "research-os-api"
version = "3.0.0"
description = "ResearchOS Unified Backend - AI Agent Orchestration & Visual Intelligence"
authors = [
    {name = "ResearchOS Team",email = "team@research-os.dev"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi[all] (>=0.116.0)",
    "uvicorn[standard] (>=0.35.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "sqlmodel (>=0.0.27,<0.1.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "psycopg[binary] (>=3.1.0,<4.0.0)",
    "python-multipart (>=0.0.20)",
    "pydantic (>=2.11.0,<3.0.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "langgraph (>=0.2.0,<3.0.0)",
    "langchain-core (>=0.3.0,<2.0.0)",
    "langchain-openai (>=0.2.0,<1.0.0)",
    "langchain (>=0.3.0,<2.0.0)",
    "langchain-community (>=0.3.0,<2.0.0)",
    "langchain-google-genai (>=2.0.0,<5.0.0)",
    "langchain-postgres (>=0.0.9,<0.1.0)",
    "pypdf (>=4.0.0,<5.0.0)",
    "tiktoken (>=0.7.0,<0.8.0)",
    "boto3 (>=1.34.0,<2.0.0)",
    "redis (>=5.0.0,<7.0.0)",
    "celery (>=5.0.0,<6.0.0)",
    "pillow (>=10.0.0,<12.0.0)",
    "numpy (>=1.24.0,<3.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "torchvision (>=0.23.0,<1.0.0)",
    "transformers (>=4.57.0,<6.0.0)",
    "qwen-vl-utils (>=0.0.14,<1.0.0)",
    "scipy (>=1.11.0,<2.0.0)",
    "jinja2 (>=3.1.0,<4.0.0)",
    "httpx (>=0.28.0,<1.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
